+++ src/Kit/DataEntry/InputWrapper/types.d.ts
@@ -8,8 +8,11 @@ export interface IKitInputWrapper extends KitHTMLAttributes<HTMLDivElement> {
     status?: '' | 'warning' | 'error';
     className?: string;
     bordered?: boolean;
+    hoverable?: boolean;
     required?: boolean;
     infoIcon?: ReactNode;
     actions?: ReactElement[];
     onInfoClick?: (e: MouseEvent<HTMLElement, MouseEvent>) => void;
+    onFocus?: (e: FocusEvent<HTMLDivElement>) => void;
+    onBlur?: (e: FocusEvent<HTMLDivElement>) => void;
 }