+++ stories/02-Design System/04-DataEntry/InputWrapper/data.ts
@@ -6,7 +6,8 @@ export const argTypes = {
         table: {
             type: {
                 summary: 'string'
-            }
+            },
+            category: 'InputWrapper'
         }
     },
     helper: {
@@ -16,7 +17,8 @@ export const argTypes = {
         table: {
             type: {
                 summary: 'string'
-            }
+            },
+            category: 'InputWrapper'
         }
     },
     required: {
@@ -27,30 +29,48 @@ export const argTypes = {
             type: {
                 summary: 'boolean'
             },
+            category: 'InputWrapper',
             defaultValue: {summary: false}
         },
         defaultValue: false
     },
     disabled: {
         name: 'disabled',
-        description: 'Whether the input is disabled',
+        description: 'Whether the content is disabled',
         control: {type: 'boolean'},
         table: {
             type: {
                 summary: 'boolean'
             },
+            category: 'InputWrapper',
             defaultValue: {summary: false}
         },
         defaultValue: false
     },
     bordered: {
         name: 'bordered',
-        description: 'Whether the input is disabled',
+        description: 'Whether the content is bordered',
         control: {type: 'boolean'},
         table: {
             type: {
                 summary: 'boolean'
             },
+            category: 'InputWrapper',
+
+            defaultValue: {summary: false}
+        },
+        defaultValue: false
+    },
+    hoverable: {
+        name: 'hoverable',
+        description: 'Whether the input border color changes on hover. Only works if `bordered` is set to true',
+        control: {type: 'boolean'},
+        table: {
+            type: {
+                summary: 'boolean'
+            },
+            category: 'InputWrapper',
+
             defaultValue: {summary: false}
         },
         defaultValue: false
@@ -63,7 +83,8 @@ export const argTypes = {
         table: {
             type: {
                 summary: "'error' | 'warning'"
-            }
+            },
+            category: 'InputWrapper'
         }
     },
     infoIcon: {
@@ -75,6 +96,8 @@ export const argTypes = {
             type: {
                 summary: 'ReactNode'
             },
+            category: 'InputWrapper',
+
             defaultValue: {summary: '<FontAwesomeIcon icon={faCircleInfo} />'}
         },
         defaultValue: false
@@ -86,6 +109,8 @@ export const argTypes = {
             type: {
                 summary: 'ReactNode[]'
             },
+            category: 'InputWrapper',
+
             defaultValue: {summary: null}
         },
         defaultValue: false
@@ -96,7 +121,8 @@ export const argTypes = {
         table: {
             type: {
                 summary: 'string'
-            }
+            },
+            category: 'InputWrapper'
         }
     },
     onInfoClick: {
@@ -106,10 +132,29 @@ export const argTypes = {
         control: {type: 'select'},
         table: {
             type: {
-                summary: '(e: MouseEvent<HTMLElement, MouseEvent>) => void;'
+                summary: '(e: MouseEvent<HTMLElement>) => void;'
             },
-            defaultValue: {summary: '-'}
-        },
-        defaultValue: () => {}
+            category: 'InputWrapper'
+        }
+    },
+    onFocus: {
+        name: 'onFocus',
+        description: 'If set, it will trigger the given function when the content is focused',
+        table: {
+            type: {
+                summary: '(e: FocusEvent<HTMLDivElement>) => void'
+            },
+            category: 'InputWrapper'
+        }
+    },
+    onBlur: {
+        name: 'onBlur',
+        description: 'If set, it will trigger the given function when focus is removed',
+        table: {
+            type: {
+                summary: '(e: FocusEvent<HTMLDivElement>) => void'
+            },
+            category: 'InputWrapper'
+        }
     }
 };